Update Friday #2: bug fixes, release prep

+++ CNU/LocationFeedItem.h
@@ -13,10 +13,10 @@
 @property(nonatomic, retain) NSString *message;
 @property(nonatomic) int minutes;
 @property(nonatomic) int crowded;
-@property(nonatomic) long time;
+@property(nonatomic) long long time;
 @property(nonatomic) bool pinned;
 @property(nonatomic, retain) NSString *detail;
 
--(id) initWithMessage:(NSString *)message withMinutes:(int)min withCrowded:(int)crow withTime:(long)t withPinned:(bool)pin withDetail:(NSString *)det;
+-(id) initWithMessage:(NSString *)message withMinutes:(int)min withCrowded:(int)crow withTime:(long long)t withPinned:(bool)pin withDetail:(NSString *)det;
 
 @end
